requirements.txt
numpy~=1.20.2
pytest~=6.2.3
pyserial~=3.5
matplotlib~=3.4.2